+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Added
 
+- Coversheet generation for performer invoices
 - System setting for performer livestreaming question
 - System setting for performer liability insurance question
 - System setting to configure acceptable back-to-back intervals
@@ -16,12 +17,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - End date to streaming program list on dashboard and streaming page
 - Scheduled task execution capability
 - Welcome email capability
+- Streamlined sign up options with authorization codes
 
 ## Changed
 
 - Improved release prep and publish scripts
 - Back-to-back interval now selectable from a list
 - Overall goal to "Community Reading Goal" so it works for all seasons
+- Changed display of "Program" on the participant side to "Age Group"
 
 ## Fixed